Pulsadores y sensores fotosensibles añadidos
Se han añadido librerías para usar los pulsadores y los fotodiodos del robot. También se ha añadido documentación de los códigos.

+++ Librerias/HAL/HAL_motores.c
@@ -1,12 +1,23 @@
+/**
+ * @file HAL_motores.c
+ * @author Jorge Ibáñez
+ * @brief Definición de la capa HAL para el uso de los motores
+ * @version 0.1
+ * @date 2025-05-03
+ * 
+ * @copyright Copyright (c) 2025
+ * 
+ */
+
 #include "HAL_motores.h"
 
-void HAL_motores_avanzar(int8_t velocidad)
+void HAL_motores_avanzar(int16_t velocidad)
 {
     motor_set_speed(MOTOR_IZQUIERDO, velocidad);
     motor_set_speed(MOTOR_DERECHO, velocidad);
 }
 
-void HAL_motores_girar(int8_t velocidad_izq, int8_t velocidad_der)
+void HAL_motores_girar(int16_t velocidad_izq, int16_t velocidad_der)
 {
     motor_girar(velocidad_izq, velocidad_der);
 }
